+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [7.9.0] - 2026-08-27
+
+_Stable release based on [7.9.0-rc.1]._
+
 ## [7.9.0-rc.1] - 2026-08-27
 
 ### Changed
@@ -1329,6 +1333,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[7.9.0]: https://https://github.com/internetguru/laravel-user/compare/v7.8.0...v7.9.0
 [7.9.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v7.8.0
 [7.8.0]: https://https://github.com/internetguru/laravel-user/compare/v7.7.3...v7.8.0
 [7.8.0-rc.1]: https://github.com/internetguru/laravel-user/releases/tag/v7.7.3